+++ test/Spec.hs
@@ -4,11 +4,13 @@ import Test.Tasty
 
 import qualified Test.Venusia.Server as Server
 import qualified Test.Venusia.MenuBuilder as MenuBuilder
+import qualified Test.Venusia.FileHandler as FileHandler
 import qualified Test.Venusia.Integration as Integration
 
 main :: IO ()
 main = defaultMain $ testGroup "Venusia"
   [ Server.tests
   , MenuBuilder.tests
+  , FileHandler.tests
   , Integration.tests
   ]